# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Asynchronous REST API for scraping real estate listings from WG-Gesucht.",
        "title": "Immogucker API",
        "contact": {
            "name": "Roman Mishyn"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Get the health status of the API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Show the status of server",
                "responses": {
                    "200": {
                        "description": "Server is up and running",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tasks": {
            "post": {
                "description": "Adds a new asynchronous scraping task to the queue based on city and max price.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a parsing task",
                "parameters": [
                    {
                        "description": "Task parameters (e.g., city, max_price, email)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Wizzerin_immogucker-go_internal_models.TaskRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Task successfully added to queue",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Retrieves the current status of a parsing task and the list of apartments if completed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task status and results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Wizzerin_immogucker-go_internal_models.TaskRequest": {
            "type": "object",
            "required": [
                "city",
                "email",
                "max_price"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "max_price": {
                    "type": "integer"
                }
            }
        }
    }
}